server/requirements.txt
flask
redis
Werkzeug
waitress
psycopg2-binary
python-dateutil